macOSにて、ランチャーのサーバ変更するとlook and feel関連のエラーが出るのを修正(ランチャーではSystemLookAndFeelを使うようにした。)
